+++ source/ssao.cpp
@@ -157,7 +157,7 @@ SSAO::createSphereNoise()
 void
 SSAO::onActivate(RenderPipeline& pipeline) except
 {
-	std::size_t width, height;	
+	std::uint32_t width, height;
 	pipeline.getWindowResolution(width, height);
 
 	_texAmbient = pipeline.createRenderTexture();
